Created GetBasketballCenter
Created a function which retrieves the position and radius of the largest circle by finding the largest circular contour and returning this as position and radius. Also added a function to trace the contour of the ball (circle) on a video by extracting each frame of the video and performing the getBasketballCenter function.